requirements.txt
Django==2.1
django-extensions==2.1.0
djangorestframework==3.8.2
freeze==1.0.10
psycopg2==2.7.4
pytz==2018.5
six==1.11.0
